Refactor references to "base tools", small UI tweaks
